==> 5.3_analyse_patches.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
os.chdir('/content/drive/My Drive/University/ProjetML/')

import pickle
import numpy as np


import os
import shutil
import itertools
import random
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import tensorflow as tf

import keras
from keras import models
from keras import backend as K
from keras.applications import imagenet_utils
from keras.preprocessing.image import ImageDataGenerator
from keras.preprocessing import image
from keras.metrics import sparse_categorical_crossentropy
from keras.optimizers import Adam
from keras.models import Sequential, Model
from keras.layers import Activation
from keras.layers.core import Dense, Flatten
from keras.layers.convolutional import *
from sklearn.metrics import confusion_matrix

with open("Data/Sean_pickles/all_patches_rescaled_array.pkl", "rb") as f:
  all_patches_rescaled_array = pickle.load(f)
with open("Data/Sean_pickles/all_patches_array.pkl", "rb") as f:
  all_patches_array = pickle.load(f)
with open("Data/Sean_pickles/all_patches_ids.pkl", "rb") as f:
  all_patches_ids = pickle.load(f)

# convert lists to np arrays
all_patches_rescaled_array = np.array(all_patches_rescaled_array)
all_patches_array = np.array(all_patches_array)
all_patches_ids = np.array(all_patches_ids)

## Manually designed models

### Simple 2D CNN Model

model = Sequential([
        Conv2D(32, (3, 3), activation='relu', input_shape=(9,9,3)),
        Flatten(), # flatten into 1-D tensor
        Dense(6, activation='softmax') #output layer that categorizes images
])
model.compile(Adam(lr=0.01), loss='sparse_categorical_crossentropy', metrics=['accuracy'])
model.summary()

model.fit(x=all_patches_array, y=all_patches_ids, batch_size=128, epochs=10, validation_split=0.2)

Test for farm optimising

model = Sequential([
        Conv2D(64, (3, 3), activation='relu', input_shape=(9,9,3)),
        Conv2D(64, (3, 3), activation='relu', padding="same"),
        keras.layers.MaxPooling2D(),

        Conv2D(32, (3, 3), activation='relu', padding="same"),
        Conv2D(32, (3, 3), activation='relu', padding="same"),
        keras.layers.MaxPooling2D(),

        Flatten(), # flatten into 1-D tensor
        Dense(6, activation='softmax') #output layer that categorizes images
])
model.compile(Adam(lr=0.01), loss='sparse_categorical_crossentropy', metrics=['accuracy'])
model.summary()
model.fit(x=all_patches_array, y=all_patches_ids, batch_size=128, epochs=10, validation_split=0.2)

### 4 layer conv

model = Sequential([
        Conv2D(32, (3, 3), activation='relu', input_shape=(9,9,3)),
        Conv2D(32, (3, 3), activation='relu', padding="same"),
        keras.layers.MaxPooling2D(),

        Conv2D(32, (3, 3), activation='relu', padding="same"),
        Conv2D(32, (3, 3), activation='relu', padding="same"),
        keras.layers.MaxPooling2D(),

        Flatten(), # flatten into 1-D tensor
        Dense(6, activation='softmax') #output layer that categorizes images
])
model.compile(Adam(lr=0.01), loss='sparse_categorical_crossentropy', metrics=['accuracy'])
model.summary()

model.fit(x=all_patches_array, y=all_patches_ids, batch_size=128, epochs=10, validation_split=0.2)

#### Comparison with rescaled:

model.fit(x=all_patches_rescaled_array, y=all_patches_ids, batch_size=128, epochs=10, validation_split=0.2)

### Comparison with Anti-overfit measures

model = Sequential([
        Conv2D(32, (3, 3), activation='relu', input_shape=(9,9,3)),
        Conv2D(32, (3, 3), activation='relu', padding="same"),
        keras.layers.MaxPooling2D(),

        Conv2D(32, (3, 3), activation='relu', padding="same"),
        Conv2D(32, (3, 3), activation='relu', padding="same"),
        keras.layers.MaxPooling2D(),

        keras.layers.BatchNormalization(),
        Flatten(), # flatten into 1-D tensor
        keras.layers.Dropout(0.5),
        Dense(6, activation='softmax') #output layer that categorizes images
])
model.compile(Adam(lr=0.0001), loss='sparse_categorical_crossentropy', metrics=['accuracy'])
model.summary()

model.fit(x=all_patches_array, y=all_patches_ids, batch_size=128, epochs=10, validation_split=0.2)

## Transfer Learning

vgg16_model = keras.applications.vgg16.VGG16(include_top=True)

vgg16_model.summary()

model = Sequential()
for layer in vgg16_model.layers:
    model.add(layer)

model.summary()

NUM_TRAINABLE_LAYERS = 10
for layer in model.layers[:-NUM_TRAINABLE_LAYERS]:
    layer.trainable = False #excludes layer from future training so weights will not be updated when we train the model on our own data

model.add(Flatten())
model.add(Dense(512, activation='relu'))
model.add(Dense(512, activation='relu'))
model.add(Dense(6, activation='softmax'))

model.compile(Adam(lr=0.0001), loss='categorical_crossentropy', metrics=['accuracy'])

model.fit(x=all_patches_array, y=all_patches_ids, batch_size=128, epochs=10, validation_split=0.2)

acc = model.evaluate(x_test, y_test)
print("Loss:", acc[0], " Accuracy:", acc[1])
